add shopping cart with and save orders to db

+++ db/migrate/20190527165809_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
     create_table :products do |t|
       t.string :title, null: false
       t.text :description
-      t.decimal :price, precision: 10, scale: 2
+      t.float :price, precision: 10, scale: 2
       t.string :image
       t.integer :orders, default: 0
       t.integer :views, default: 0
